separating validatePdf behavior from interactor
creating use-case helper function
