+++ shared/src/business/useCases/filePetitionInteractor.js
@@ -44,8 +44,8 @@ exports.filePetitionInteractor = async (
         key,
       });
     await applicationContext
-      .getUseCases()
-      .validatePdfInteractor(applicationContext, {
+      .getUseCaseHelpers()
+      .validatePdf(applicationContext, {
         key,
       });
