+++ shared/src/business/useCases/uploadDocumentAndMakeSafeInteractor.js
@@ -23,11 +23,9 @@ exports.uploadDocumentAndMakeSafeInteractor = async (
     .getStatusOfVirusScanInteractor(applicationContext, {
       key: uploadedKey,
     });
-  await applicationContext
-    .getUseCases()
-    .validatePdfInteractor(applicationContext, {
-      key: uploadedKey,
-    });
+  await applicationContext.getUseCaseHelpers().validatePdf(applicationContext, {
+    key: uploadedKey,
+  });
 
   return uploadedKey;
 };
